test: created new fixtures and progress tests
