+++ test/fixtures/fixtures.dart
@@ -1,0 +1,2 @@
+export 'mocks/mocks.dart';
+export 'user_fixtures.dart';
