+++ test/fixtures/mocks/mock_app_logger.dart
@@ -1,0 +1,4 @@
+import 'package:good_app/app/core/logger/logger.dart';
+import 'package:mocktail/mocktail.dart';
+
+class MockAppLogger extends Mock implements AppLogger {}
