+++ test/fixtures/mocks/mock_login_data_source.dart
@@ -1,0 +1,4 @@
+import 'package:good_app/features/features.dart';
+import 'package:mocktail/mocktail.dart';
+
+class MockLoginDataSource extends Mock implements LoginDataSource {}
